+++ src/App.tsx
@@ -1,4 +1,4 @@
-import { Routes, Route } from 'react-router-dom'
+import { Switch, Route } from 'wouter'
 import Layout from './components/Layout'
 import TaskList from './pages/TaskList'
 import AddTask from './pages/AddTask'
@@ -6,13 +6,13 @@ import Stats from './pages/Stats'
 
 function App() {
   return (
-    <Routes>
-      <Route path="/" element={<Layout />}>
-        <Route index element={<TaskList />} />
-        <Route path="add" element={<AddTask />} />
-        <Route path="stats" element={<Stats />} />
-      </Route>
-    </Routes>
+    <Layout>
+      <Switch>
+        <Route path="/" component={TaskList} />
+        <Route path="/add" component={AddTask} />
+        <Route path="/stats" component={Stats} />
+      </Switch>
+    </Layout>
   )
 }