+++ src/pages/Stats.tsx
@@ -1,5 +1,5 @@
 import { useState, useMemo } from 'react'
-import { Link } from 'react-router-dom'
+import { Link } from 'wouter'
 import { useTasks } from '../context/TaskContext'
 
 const monthNames = [